# frontend/openapi.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "ErrorResponse": {
                "properties": { "error": { "type": "string" }, "message": { "type": "string" } },
                "required": ["error", "message"],
                "type": "object",
                "additionalProperties": false
            },
...
            "SuccessResponse": {
                "properties": { "message": { "type": "string" } },
                "required": ["message"],
                "type": "object",
                "additionalProperties": false
            },
...
            "Schedule": {
                "properties": {
                    "id": { "type": "string" },
                    "date": { "type": "string", "format": "date-time" },
                    "relationshipId": { "type": "string" },
                    "startTime": { "type": "string" },
                    "endTime": { "type": "string" }
                },
                "required": ["id", "date", "relationshipId", "startTime", "endTime"],
                "type": "object",
                "additionalProperties": false
            },
            "ScheduleRequest": {
                "properties": {
                    "date": { "type": "string", "format": "date-time" },
                    "relationshipId": { "type": "string" },
                    "startTime": { "type": "string" },
                    "endTime": { "type": "string" }
                },
                "required": ["date", "relationshipId", "startTime", "endTime"],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": { "jwt": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
    },
    "info": { "title": "activity-log-app-backend", "version": "0.0.0", "license": { "name": "ISC" }, "contact": {} },
    "paths": {
...
        "/schedules": {
            "post": {
                "operationId": "CreateSchedule",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/Schedule" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing fields",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Schedules"],
                "security": [{ "jwt": [] }],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleRequest" } } }
                }
            },
            "get": {
                "operationId": "GetSchedules",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "items": { "$ref": "#/components/schemas/Schedule" }, "type": "array" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Schedules"],
                "security": [{ "jwt": [] }],
                "parameters": []
            }
        },
        "/schedules/{id}": {
            "get": {
                "operationId": "GetSchedule",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/Schedule" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Schedule not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Schedules"],
                "security": [{ "jwt": [] }],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }]
            },
            "put": {
                "operationId": "UpdateSchedule",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/SuccessResponse" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing fields",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "404": {
                        "description": "Schedule not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Schedules"],
                "security": [{ "jwt": [] }],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleRequest" } } }
                }
            },
            "delete": {
                "operationId": "DeleteSchedule",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/SuccessResponse" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Schedule not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Schedules"],
                "security": [{ "jwt": [] }],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }]
            }
        },
        "/schedules/caregiver/{caregiverId}": {
            "get": {
                "operationId": "GetSchedulesForCaregiver",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "items": { "$ref": "#/components/schemas/Schedule" }, "type": "array" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Schedules"],
                "security": [{ "jwt": [] }],
                "parameters": [
                    { "in": "path", "name": "caregiverId", "required": true, "schema": { "type": "string" } }
                ]
            }
        },
        "/schedules/recipient/{recipientId}": {
            "get": {
                "operationId": "GetSchedulesForRecipient",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "items": { "$ref": "#/components/schemas/Schedule" }, "type": "array" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Schedules"],
                "security": [{ "jwt": [] }],
                "parameters": [
                    { "in": "path", "name": "recipientId", "required": true, "schema": { "type": "string" } }
                ]
            }
        },
        "/schedules/caregiver/{caregiverId}/recipient/{recipientId}": {
            "get": {
                "operationId": "GetSchedulesForCaregiverAndRecipient",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "items": { "$ref": "#/components/schemas/Schedule" }, "type": "array" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Schedules"],
                "security": [{ "jwt": [] }],
                "parameters": [
                    { "in": "path", "name": "caregiverId", "required": true, "schema": { "type": "string" } },
                    { "in": "path", "name": "recipientId", "required": true, "schema": { "type": "string" } }
                ]
            }
        },
...
    },
    "servers": [{ "url": "/" }]
}